bug fixes, added testing files
